+++ src/test/exception.test.ts
@@ -1,11 +1,11 @@
-import {numberToChineseWords} from '../main/index';
+import {convertNumber} from '../main/index';
 
 test.each([[-10], [-1], [10.6], [50.55]])(
   'Not Supported error',
   (input: number) => {
     const errmsg = 'Not Supported';
     expect(() => {
-      numberToChineseWords(input);
+      convertNumber(input);
     }).toThrowError(errmsg);
   }
 );