+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - change version to `1.0.0`
 - update readme with sample outputs
+- change function name to `convertNumber`
 
 ## `0.0.6` - 10 June 2022